Add files via upload
In this project, we aim to analyze iPhone purchase records stored in a CSV file using the K-Nearest Neighbors (KNN) algorithm implemented in Python. The dataset contains information such as customer age, income, gender, and whether they purchased an iPhone or not. By leveraging the KNN algorithm, we intend to build a classification model that can predict whether a customer is likely to purchase an iPhone based on their demographic attributes.